+++ index.html
@@ -249,7 +249,7 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="method"></script>
 </body>
 
 </html>
