+++ js/util.js
@@ -1,0 +1,9 @@
+// Возвращает произвольное целое число из диапазона min..max
+const getRandomNumber = (min, max) =>
+  Math.floor(Math.random() * (max - min + 1) + min);
+
+// Возвращант произвольный элемент массива
+const getRandomArrayElement = (array) =>
+  array[getRandomNumber(0, array.length - 1)];
+
+export { getRandomNumber, getRandomArrayElement };
